// src/common/proto/project.pb.ts
export enum StatusType {
  START = 0,
  IN_PROGRESS = 1,
  FINISHED = 2,
  DELETED = 3,
  UNRECOGNIZED = -1,
}

export enum LevelType {
  LOW = 0,
  HIGH = 1,
  UNRECOGNIZED = -1,
}

export enum UnityType {
  VB = 0,
  M2 = 1,
  UNID = 2,
  UNRECOGNIZED = -1,
}

/**
 * Create Project
 * Request
 */
export interface ProjectCreateRequest {
  startDate: string;
  expectedFinishedDate: string;
  responsible: string;
  address: Address | undefined;
  tasks: Task[];
  status: StatusType;
}

/** Response */
export interface ProjectCreateResponse {
  status: number;
  error: string;
}

/** FindAll Project */
export interface ProjectFindAllRequest {}

export interface ProjectFindAllResponse {
  status: number;
  error: string;
  data: FindAllData[];
}

export interface FindAllData {
  status: string;
  startDate: number;
  expectedFinishedDate: number;
  responsible: string;
  address: Address | undefined;
  id: string;
}

export interface ProjectFindOneRequest {
  id: string;
}

export interface ProjectFindOneResponse {
  status: number;
  error: string;
  data: FindOneData | undefined;
}

export interface FindOneData {
  status: string;
  startDate: number;
  expectedFinishedDate: number;
  responsible: string;
  address: Address | undefined;
  id: string;
}

/**
 * Update Project
 * Request
 */
export interface ProjectUpdateRequest {
  id: string;
  data: UpdateProjectData | undefined;
}

/** Response */
export interface ProjectUpdateResponse {
  status: number;
  error: string;
  data: UpdateData | undefined;
}

export interface UpdateData {
  status: string;
  startDate: number;
  expectedFinishedDate: number;
  responsible: string;
  address: Address | undefined;
  id: string;
}

export interface UpdateProjectData {
  status?: StatusType | undefined;
  startDate?: string | undefined;
  expectedFinishedDate?: string | undefined;
  responsible?: string | undefined;
}

/**
 * Remove Project
 * Request
 */
export interface ProjectRemoveRequest {
  id: string;
}

/** Response */
export interface ProjectRemoveResponse {
  status: number;
  error: string;
  data: RemoveData | undefined;
}

export interface RemoveData {}

/**
 * FindAllTaskOfProjectRequest
 * Request
 */
export interface FindAllTaskOfProjectRequest {
  id: string;
}

/** Response */
export interface FindAllTaskOfProjectResponse {
  status: number;
  error: string;
  data: FindAllTaskOfProjectData[];
}

export interface FindAllTaskOfProjectData {
  category: string;
  activity: string;
  noiseLevel: LevelType;
  dirtLevel: LevelType;
  description: string;
  unity: UnityType;
  possibleSkills: PossibleSkills[];
}

/**
 * FieldsOverwriters
 * Request
 */
export interface FieldsOverridesRequest {
  projectId: string;
  taskId: string;
  data: FieldsOverrides | undefined;
}

/** Response */
export interface FieldsOverridesResponse {
  status: number;
  error: string;
  data: FieldsOverrides | undefined;
}

export interface FieldsOverrides {
  category?: string | undefined;
  activity?: string | undefined;
  noiseLevel?: LevelType | undefined;
  dirtLevel?: LevelType | undefined;
  description?: string | undefined;
  unity?: UnityType | undefined;
}

export interface Address {
  zipCode: string;
  street: string;
  number: string;
  neighborhood: string;
  city: string;
  state: string;
}

export interface Task {
  category: string;
  activity: string;
  noiseLevel: LevelType;
  dirtLevel: LevelType;
  description: string;
  unity: UnityType;
  possibleSkills: PossibleSkills[];
  laborers: string[];
}

export interface PossibleSkills {
  skillId: string;
  requiredSkillLevel: number;
}